requirements.txt
ultralytics==8.3.3
flask==3.1.1
sqlalchemy==2.0.45
opencv-python-headless==4.11.0.76
torch==2.1.2
torchvision==0.16.1
scipy==1.12.1
pyyaml==6.0.2